Added convenience methods to RepositoryFactories static class, also changed numeric type on MemoryRepository

+++ CSF.Patterns.DDD/Data/Memory/MemoryRepository.cs
@@ -33,7 +33,7 @@ namespace CSF.Patterns.DDD.Data.Memory
     #region fields
     
     private Dictionary<IIdentity, TEntity> _backingStore;
-    private uint _autoIncrement;
+    private ulong _autoIncrement;
     
     #endregion
